pneumo-ai-pipeline/models/requirements.txt
# Core ML packages
torch==1.13.1
torchvision==0.14.1
torchaudio==0.13.1  # PyTorch 오디오 관련 기능용
numpy==1.26.4
pillow==11.3.0

# MLflow and dependencies
mlflow==2.8.1
protobuf>=3.12.0,<5.0.0
setuptools>=65.5.1
importlib-metadata>=6.0.0

# AWS packages
boto3>=1.39.6  # S3 접근용
sagemaker>=2.248.1  # SageMaker SDK

# Utilities
tqdm==4.67.1
PyYAML==6.0.2
python-dotenv==1.1.1

# Additional dependencies
torchxrayvision==1.3.2

transformers==4.40.0
kornia>=0.6.0